test(charts): consolidate the sized-chart test helper into chart-layout
main and this branch each landed a copy of the same recharts-sizing test
helper under a different name, so the rebase left two of them side by side.
Keep `chart-layout` (already imported by the RadarChart tests) and fold in
the one thing `sized-chart` did better: scoping the faked rect to the
responsive container, so the legend is measured normally instead of claiming
the whole plot area.

`giveTheChartASize` is now a no-op when already installed. The curve loops in
the Line/Area tests used to restore between iterations; without the guard a
second call would capture this wrapper as its own "original" and leave it
installed after the restore.
